backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pandas
requests
google-cloud-storage
python-magic
pytest
httpx
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
